LC-maximize-happiness-of-selected-children [Runtime: 141 ms; Memory: 107.8 MB]

+++ array/maximize-happiness-of-selected-children.cpp
@@ -3,8 +3,8 @@ public:
     long long maximumHappinessSum(vector<int>& happiness, int k) {
         sort(happiness.begin(), happiness.end());
 
-        int ans=0;
-        int t=0;
+        long long ans=0;
+        long long t=0;
         int n=happiness.size()-1;
         for(int i=0; i<k;i++){
            if(t<=happiness[n]) ans+=happiness[n--]-t;
